+++ app/src/main/java/com/naver/maps/map/demo/kotlin/overlay/OverlayMinMaxZoomActivity.kt
@@ -15,28 +15,27 @@
  */
 package com.naver.maps.map.demo.kotlin.overlay
 
-import android.graphics.Color
-import android.graphics.PointF
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.view.MenuItem
 import com.naver.maps.geometry.LatLng
+import com.naver.maps.map.CameraAnimation
 import com.naver.maps.map.CameraPosition
+import com.naver.maps.map.CameraUpdate
 import com.naver.maps.map.MapFragment
 import com.naver.maps.map.NaverMap
 import com.naver.maps.map.NaverMapOptions
 import com.naver.maps.map.OnMapReadyCallback
 import com.naver.maps.map.demo.R
-import com.naver.maps.map.overlay.Align
 import com.naver.maps.map.overlay.Marker
-import com.naver.maps.map.overlay.OverlayImage
 import com.naver.maps.map.util.MarkerIcons
+import kotlinx.android.synthetic.main.activity_overlay_min_max_zoom.*
 
-class MarkerActivity : AppCompatActivity(), OnMapReadyCallback {
+class OverlayMinMaxZoomActivity : AppCompatActivity(), OnMapReadyCallback {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        setContentView(R.layout.activity_map_fragment)
+        setContentView(R.layout.activity_overlay_min_max_zoom)
 
         supportActionBar?.let {
             it.setDisplayHomeAsUpEnabled(true)
@@ -45,9 +44,7 @@ class MarkerActivity : AppCompatActivity(), OnMapReadyCallback {
 
         val mapFragment = supportFragmentManager.findFragmentById(R.id.map_fragment) as MapFragment?
                 ?: run {
-                    val options = NaverMapOptions().camera(CameraPosition(
-                            NaverMap.DEFAULT_CAMERA_POSITION.target, NaverMap.DEFAULT_CAMERA_POSITION.zoom, 30.0, 45.0))
-                    MapFragment.newInstance(options).also {
+                    MapFragment.newInstance(NaverMapOptions().camera(DEFAULT_CAMERA_POSITION)).also {
                         supportFragmentManager.beginTransaction().add(R.id.map_fragment, it).commit()
                     }
                 }
@@ -64,61 +61,59 @@ class MarkerActivity : AppCompatActivity(), OnMapReadyCallback {
 
     override fun onMapReady(naverMap: NaverMap) {
         Marker().apply {
-            position = LatLng(37.5666102, 126.9783881)
+            icon = MarkerIcons.GREEN
+            position = LatLng(37.56713851901027, 126.97891430703686)
+            captionText = getString(R.string.caption_marker_1)
+            minZoom = 15.0
+            isMinZoomInclusive = true
+            maxZoom = 17.0
+            isMaxZoomInclusive = true
             map = naverMap
         }
 
         Marker().apply {
-            position = LatLng(37.57000, 126.97618)
-            icon = MarkerIcons.BLACK
-            angle = 315f
+            icon = MarkerIcons.BLUE
+            position = LatLng(37.56713851901027, 126.97786189296272)
+            captionText = getString(R.string.caption_marker_2)
+            minZoom = 15.0
+            isMinZoomInclusive = false
+            maxZoom = 17.0
+            isMaxZoomInclusive = true
             map = naverMap
         }
 
         Marker().apply {
-            position = LatLng(37.57145, 126.98191)
-            icon = OverlayImage.fromResource(R.drawable.ic_info_black_24dp)
-            width = resources.getDimensionPixelSize(R.dimen.marker_size)
-            height = resources.getDimensionPixelSize(R.dimen.marker_size)
-            isFlat = true
-            angle = 90f
+            icon = MarkerIcons.RED
+            position = LatLng(37.566081877242425, 126.97891430703686)
+            captionText = getString(R.string.caption_marker_3)
+            minZoom = 15.0
+            isMinZoomInclusive = true
+            maxZoom = 17.0
+            isMaxZoomInclusive = false
             map = naverMap
         }
 
         Marker().apply {
-            position = LatLng(37.56768, 126.98602)
-            icon = OverlayImage.fromResource(R.drawable.marker_right_bottom)
-            anchor = PointF(1f, 1f)
-            angle = 90f
-            map = naverMap
-        }
-
-        Marker().apply {
-            position = LatLng(37.56436, 126.97499)
             icon = MarkerIcons.YELLOW
-            captionMinZoom = 12.0
-            setCaptionAligns(Align.Left)
-            captionText = getString(R.string.marker_caption_1)
+            position = LatLng(37.566081877242425, 126.97786189296272)
+            captionText = getString(R.string.caption_marker_4)
+            minZoom = 15.0
+            maxZoom = 17.0
+            isMinZoomInclusive = false
+            isMaxZoomInclusive = false
             map = naverMap
         }
 
-        Marker().apply {
-            position = LatLng(37.56138, 126.97970)
-            icon = MarkerIcons.PINK
-            captionTextSize = 14f
-            captionText = getString(R.string.marker_caption_2)
-            subCaptionTextSize = 10f
-            subCaptionColor = Color.GRAY
-            subCaptionText = getString(R.string.marker_sub_caption_2)
-            map = naverMap
+        naverMap.addOnCameraChangeListener { _, _ ->
+            zoom.text = getString(R.string.format_zoom, naverMap.cameraPosition.zoom)
         }
 
-        Marker().apply {
-            position = LatLng(37.56500, 126.9783881)
-            icon = MarkerIcons.BLACK
-            iconTintColor = Color.RED
-            alpha = 0.5f
-            map = naverMap
+        fab.setOnClickListener {
+            naverMap.moveCamera(CameraUpdate.toCameraPosition(DEFAULT_CAMERA_POSITION).animate(CameraAnimation.Easing))
         }
+    }
+
+    companion object {
+        private val DEFAULT_CAMERA_POSITION = CameraPosition(NaverMap.DEFAULT_CAMERA_POSITION.target, 16.0)
     }
 }